+++ lib/CMakeLists.txt
@@ -16,5 +16,6 @@ add_subdirectory_ifdef(CONFIG_LTE_LINK_CONTROL lte_link_control)
 add_subdirectory_ifdef(CONFIG_MODEM_INFO modem_info)
 add_subdirectory_ifdef(CONFIG_PDN_MANAGEMENT pdn_management)
 add_subdirectory_ifdef(CONFIG_MULTITHREADING_LOCK multithreading_lock)
+add_subdirectory_ifdef(CONFIG_FPROTECT fprotect)
 add_subdirectory_ifdef(CONFIG_RESET_ON_FATAL_ERROR fatal_error)
 add_subdirectory_ifdef(CONFIG_MODEM_KEY_MGMT modem_key_mgmt)
