+++ lib/fprotect/fprotect_bprot.c
