+++ lib/fprotect/fprotect_spu.c
